Перенес все в объект
